+++ flowtype/commands/__init__.py
@@ -1,9 +1,11 @@
 from .add_pragma import FlowtypeAddPragma
 from .check_contents import FlowtypeCheckContents
 from .goto_definition import FlowtypeGotoDefinition
+from .view_type import FlowtypeViewType
 
 __all__ = [
     'FlowtypeAddPragma',
     'FlowtypeCheckContents',
     'FlowtypeGotoDefinition',
+    'FlowtypeViewType',
 ]